GRU-183 consolidate vocabulary runtime
